BRILLIAN API CACHE/requirements.txt
Django==4.2.7
psycopg2==2.9.9
django-redis==5.4.0
djangorestframework==3.14.0